Additional fixes to integer types to use int64_t and uint64_t.  These changes weren't needed for mac where I made the original changes, but showed up on peloton and BG/Q.
Code now builds on seq, aztec, and mac.



git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@999 baecf98e-1e81-4516-8920-d4c889e67ea4

+++ src/BucketOfBits.cc
@@ -64,7 +64,7 @@ void BucketOfBits::Record::getValue(unsigned fieldIndex, int& value) const
    value = strtol(rawFields_[fieldIndex].c_str(), NULL, 10);
 }
 
-void BucketOfBits::Record::getValue(unsigned fieldIndex, unsigned long long& value) const
+void BucketOfBits::Record::getValue(unsigned fieldIndex, uint64_t& value) const
 {
    assert(fieldTypes_[fieldIndex] == intType);
    value = strtoull(rawFields_[fieldIndex].c_str(), NULL, 10);
