+++ src/BucketOfBits.hh
@@ -3,6 +3,7 @@
 
 #include <string>
 #include <vector>
+#include <stdint.h>
 
 class BucketOfBits
 {
@@ -18,7 +19,7 @@ class BucketOfBits
    
       void getValue(unsigned fieldIndex, double& value) const;
       void getValue(unsigned fieldIndex, int& value) const;
-      void getValue(unsigned fieldIndex, unsigned long long& value) const;
+      void getValue(unsigned fieldIndex, uint64_t& value) const;
       void getValue(unsigned fieldIndex, std::string& value) const;
       
     private:
